requirements.txt
python-gitlab>=5.3.0
colorama>=0.4.6
types-colorama>=0.4.15.20240311

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.11.1
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
coverage>=7.2.7